Update interface Piece, changing the return type of method getValue from int type to float

+++ src/main/java/it/unibo/xiangqi/model/api/Piece.java
@@ -71,7 +71,7 @@ public interface Piece {
      *
      * @return the piece value
      */
-    public int getValue();
+    public float getValue();
 
     /**
     * Updates the strategic value of this piece.
